wip: eliminate old live session graphql types
