+++ packages/graphql/src/index.ts
@@ -10,7 +10,6 @@ import './schema/groupActivity.js'
 import './schema/microLearning.js'
 import './schema/participant.js'
 import './schema/question.js'
-import './schema/session.js'
 import './schema/user.js'
 
 import './schema/mutation.js'
