+++ packages/graphql/src/schema/liveQuiz.ts
@@ -199,4 +199,21 @@ export const ElementBlock = ElementBlockRef.implement({
     }),
   }),
 })
+
+export interface IRunningLiveQuizSummary {
+  numOfResponses: number
+  numOfFeedbacks: number
+  numOfConfusionFeedbacks: number
+  numOfLeaderboardEntries: number
+}
+export const RunningLiveQuizSummaryRef =
+  builder.objectRef<IRunningLiveQuizSummary>('RunningLiveQuizSummary')
+export const RunningLiveQuizSummary = RunningLiveQuizSummaryRef.implement({
+  fields: (t) => ({
+    numOfResponses: t.exposeInt('numOfResponses'),
+    numOfFeedbacks: t.exposeInt('numOfFeedbacks'),
+    numOfConfusionFeedbacks: t.exposeInt('numOfConfusionFeedbacks'),
+    numOfLeaderboardEntries: t.exposeInt('numOfLeaderboardEntries'),
+  }),
+})
 // #endregion
